Add version 16 configuration format to ls_extstorage.ino, add microlinn uninstall path.
Version 16 is the latest mainline Linnstrument configuration format.  To distingusih MicroLinn,
we add 128 to the mainline version it's based on.  (Current Microlinn is therefore 144.)

We add the mainline version 16 format to ls_extstorage.ino so we can migrate to and from it.

We add a path for smoothly upgrading version 16 to MicroLinn.

For "uninstalling" MicroLinn, we just revert the version 144 format to version 16 before sending
it out on the serial port (when requested by the updater tool).