Set MICROLINN_VERSION_OFFSET to 56 rather than 128, do "uninstall" in-place to avoid having to allocate.
Using 128 was problematic, as the Linnstrument update utility interprets values over 127 as negative
version numbers, and is unable to negotiate the settings transfer correctly.

The in-place uninstall is done because the configuration is rather large.  Allocating large chunks of
memory on the stack or on the heap is problematic, and we weren't able to get either to work when we
tried it.